implements Home Screen View Model class
